refactor(search): move channel name resolver to comms hex outbound
The previous attempt placed batch_resolve_channel_names in
comms_db_client, but that crate is the raw-query layer and shouldn't
host domain logic. Move the function to comms::outbound::postgres
where it can call the canonical comms::domain::models::channel_name
resolver directly. call and search_service now depend on comms with
the outbound feature; the call repo's single-channel resolve_channel_name
delegates to the batch function so there is one source of truth.
