--- rust/cloud-storage/.sqlx/query-1ca2dd639722179fe276d33b4b4d57033fdf65037fc7dd1fedc374463ccd7096.json
@@ -1,34 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            SELECT u.id as user_profile_id, mui.first_name, mui.last_name\n            FROM macro_user_info mui\n            JOIN \"User\" u ON mui.macro_user_id = u.macro_user_id\n            WHERE u.id = ANY($1)\n            ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "user_profile_id",
-        "type_info": "Text"
-      },
-      {
-        "ordinal": 1,
-        "name": "first_name",
-        "type_info": "Text"
-      },
-      {
-        "ordinal": 2,
-        "name": "last_name",
-        "type_info": "Text"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "TextArray"
-      ]
-    },
-    "nullable": [
-      false,
-      true,
-      true
-    ]
-  },
-  "hash": "1ca2dd639722179fe276d33b4b4d57033fdf65037fc7dd1fedc374463ccd7096"
-}
