+++ rust/cloud-storage/.sqlx/query-4103298cf62438b69d534ac4092423b210d7bcf2202ca38480fd7ce026977166.json
@@ -1,6 +1,6 @@
 {
   "db_name": "PostgreSQL",
-  "query": "\n            SELECT channel_id, user_id\n            FROM comms_channel_participants\n            WHERE channel_id = ANY($1) AND left_at IS NULL\n            ",
+  "query": "\n        SELECT channel_id, user_id\n        FROM comms_channel_participants\n        WHERE channel_id = ANY($1) AND left_at IS NULL\n        ",
   "describe": {
     "columns": [
       {
@@ -24,5 +24,5 @@
       false
     ]
   },
-  "hash": "a42ed7aaed8d1c0d84fe66d9ed2dc61ff75417b2134173c46cd1cdc171b8f726"
+  "hash": "4103298cf62438b69d534ac4092423b210d7bcf2202ca38480fd7ce026977166"
 }
