--- rust/cloud-storage/.sqlx/query-767162a3de69fd6df368e97dbf8ffdca791dfc0f1557adc7203b89ee8082251d.json
@@ -1,34 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n        SELECT id, name, channel_type::text as \"channel_type!\"\n        FROM comms_channels\n        WHERE id = ANY($1)\n        ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "id",
-        "type_info": "Uuid"
-      },
-      {
-        "ordinal": 1,
-        "name": "name",
-        "type_info": "Varchar"
-      },
-      {
-        "ordinal": 2,
-        "name": "channel_type!",
-        "type_info": "Text"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "UuidArray"
-      ]
-    },
-    "nullable": [
-      false,
-      true,
-      null
-    ]
-  },
-  "hash": "767162a3de69fd6df368e97dbf8ffdca791dfc0f1557adc7203b89ee8082251d"
-}
