--- rust/cloud-storage/.sqlx/query-f4f6aabb3ebbf6f8d2361b1cf8a63948bd3b0becdfdd21d780a08ffa05b2b963.json
@@ -1,28 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            SELECT name, channel_type::text as \"channel_type!\"\n            FROM comms_channels\n            WHERE id = $1\n            ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "name",
-        "type_info": "Varchar"
-      },
-      {
-        "ordinal": 1,
-        "name": "channel_type!",
-        "type_info": "Text"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "Uuid"
-      ]
-    },
-    "nullable": [
-      true,
-      null
-    ]
-  },
-  "hash": "f4f6aabb3ebbf6f8d2361b1cf8a63948bd3b0becdfdd21d780a08ffa05b2b963"
-}
