+++ rust/cloud-storage/comms/src/outbound/postgres.rs
@@ -1,2 +1,3 @@
+pub mod channel_name;
 pub mod comms_repo;
 pub mod user_repo;
